fixed the environment updates
