+++ frontend/src/model/MainJsonContext.js
@@ -7,53 +7,54 @@ const MainJsonContext = createContext();
 export const useMainJson = () => useContext(MainJsonContext);
 
 export const MainJsonProvider = ({ children }) => {
-  const [mainJson, setJson] = useState(new SimulationConfigurationModel());
+  const [mainJson, setMainJson] = useState(new SimulationConfigurationModel());
+  const [envJson, setEnvJson] = useState(mainJson.environment);
 
-  const setMainJson = (envJson, id) => {
-    if(id == "environment" && mainJson.Drones && mainJson.Drones[0].X != envJson.Origin.Latitude) {
-      setJson(prevState => ({
-        ...prevState,
-        Drones:null,
-      }))
-    }
-    setJson(prevState => ({
-      ...prevState,
-      [id]: envJson
-    }))
-  }
+  // const setMainJson = (envJson, id) => {
+  //   if(id == "environment" && mainJson.Drones && mainJson.Drones[0].X != envJson.Origin.Latitude) {
+  //     setJson(prevState => ({
+  //       ...prevState,
+  //       Drones:null,
+  //     }))
+  //   }
+  //   setJson(prevState => ({
+  //     ...prevState,
+  //     [id]: envJson
+  //   }))
+  // }
 
-  const setDroneJson = (json, index) => {
-    setJson(prevState => {
-        const updatedDrones = prevState.Drones?.map((drone, idx) => {
-            if (idx === index) {
-                return {...drone, ...json};
-            }
-            return drone;
-        });
+  // const setDroneJson = (json, index) => {
+  //   setJson(prevState => {
+  //       const updatedDrones = prevState.Drones?.map((drone, idx) => {
+  //           if (idx === index) {
+  //               return {...drone, ...json};
+  //           }
+  //           return drone;
+  //       });
 
-        return {...prevState, Drones: updatedDrones};
-    });
-  }
+  //       return {...prevState, Drones: updatedDrones};
+  //   });
+  // }
 
-  const setDroneLocation = (droneInx, longitude, latitude) => {
-    setJson(prevState => {
-      const newDrones = prevState.Drones?.map((drone, index) => {
-        if (index === droneInx) {
-          return {
-            ...drone,
-            X: latitude,
-            Y: longitude
-          };
-        }
-        return drone;
-      });
+  // const setDroneLocation = (droneInx, longitude, latitude) => {
+  //   setJson(prevState => {
+  //     const newDrones = prevState.Drones?.map((drone, index) => {
+  //       if (index === droneInx) {
+  //         return {
+  //           ...drone,
+  //           X: latitude,
+  //           Y: longitude
+  //         };
+  //       }
+  //       return drone;
+  //     });
 
-      return {...prevState, Drones: newDrones};
-    });
-  };
+  //     return {...prevState, Drones: newDrones};
+  //   });
+  // };
 
   return (
-    <MainJsonContext.Provider value={{ mainJson, setJson, setMainJson, setDroneLocation, setDroneJson }}>
+    <MainJsonContext.Provider value={{ mainJson, setMainJson, envJson, setEnvJson }}>
       {children}
     </MainJsonContext.Provider>
   );
